Refactor all firestore actions into common actions (#719)
